[ADD] MainActivity Viewholder Connect
list_item.xml correct, recyclerView connect, dummy data insert

+++ app/src/main/java/com/example/molaschoolproject/SchoolProfiles.kt
@@ -1,3 +1,3 @@
 package com.example.molaschoolproject
 
-class SchoolProfiles(val schoolName:String, val point:Float, val count:Int, val schoolType:String, val profile:Int)
+class SchoolProfiles(val schoolName:String, val point:Double, val count:Int, val schoolType:String, val profile:Int)
